Trigger restarts for nova-api-metadata
The neutron-gateway charm operates a nova-api-metadata service
for instance access to Nova metadata; this needs to be restarted
after a database initialization as if the conductors are not
responding, this service will shut itself down.

Set a remote_restart nonce on the quantum-network-service
relation post database migration to trigger remote restarts.

Change-Id: I469d4119dd95cc51378d7ca2e3ef736d94c12226
Closes-Bug: 1547122