Run flavor migrate on upgrades.
Force upgrades to Liberty to step through Kilo and force Kilo upgrades to
migrate flavor information. The flavor migration needs to be run with a
Kilo tool set as the flavor migration option appears to have been removed
in Liberty . The charm now run the flavor migration if the current
deployment is Kilo and the user is migrating to a higher release.

When an instance is created on Icehouse the flavour ID is stored in a table
called instance_system_metadata. Kilo (or maybe Juno) creates a new table
in the nova database called instance_extra. For new instances created after
upgrading to Kilo the flavour information goes into the instance_extra
table not instance_system_metadata. It is a prerequisite for upgrading to
Liberty that this flavour data has been moved from instance_system_metadata
to instance_extra. At upgrade at this point will fail because the instance
created when the system was on Icehouse has its flavour information in the
wrong place.

Change-Id: If486372b05d165bd17fa9466b435f22088532d89
Closes-Bug: #1580698