Restart neutron-gateway services on amqp complete
The nova-api-metadata service which runs in the neutron-gateway
charm units will shutdown if it does not receive a response from
a nova-conductor within a short timeout period; this can happen
if the nova-conductor is not yet wired to the messaging bus.

Trigger a remote restart using the existing relation api when
the amqp relation is complete, and nova-conductor services should
be (shortly) up and running.

Change-Id: I47d242a6e4a11d30400b9b787ace752a472b2d1e
Closes-Bug: 1547122
(cherry picked from commit 928644532b74d9bcb442d1fbb85eef7b03662934)