Support multi-phase access-network configuration
The introduction of the nova_api database in Mitaka broke support
for Percona Cluster deployed using the access-network configuration.

Ensure that a None return value from SharedDBContext is passed up
the call chain, indicating that the relation is currently incomplete
for the nova_api database; a later data set will complete the
relation.

Change-Id: I8b6a95832732ea7dd7d0f25ec63e403797359edb
Closes-Bug: 1618321