Get glance v1 client for icehouse.
Previously the image virt type was qemu and the compute node virt
type was kvm. This works for deployments prior to rocky but in
rocky this causes the image type filter to return no valid hosts.
An update to charmhelpers has removed the default behaviour of
setting the virt type to 'qemu' by default.

Due to a bug in icehouse updating glance image properties using
the v2 api fails (See Bug #1371559) so for icehouse deploys
get a v1 client.

Change-Id: I20548118092a4480f37c7ab7e9d60e72e299989b