Disable BuildFailureWeigher
Disable the BuildFailureWeigher used when weighting hosts during
instance scheduling. A single build failure will result in a
-1000000 weighting which effectively excludes the hypervisor
from the scheduling decision.

A bad image can result in build failures resulting in a heavy
load on hypervisors which have not had a build failure with
those that have effectively being ignored; the build failure
count will be reset on a successful build but due to the high
weighting this won't happen until all resources on known good
hypervisors have been completely consumed.

Change-Id: I4d4367ef20e2a20aee1e26d4a0ec69cad2ac69d6
Closes-Bug: 1818239