Refactor update_ssh_keys_and_notify_compute_units()
This patchset refactors the update_ssh_keys_and_notify_compute_units()
into two separate halves: one is "update keys and hosts from a specified
unit on a relation (or the current relation/unit for the hook)" and the
other is update the relation_set data for the relation with the found
keys/hosts.

This is a precusor patch to reducing the number of dns queries and
setting of relation data, and to eliminate repeated operations on the
same data/result (which currently happens).

Change-Id: I45bc9a889968796572a61c199ac25d543c064670
Related-Bug: 1833420