Refactor region notification code to not need unit
The (already) refactored region notification code checked for the
'region' value in the remote unit, despite the nova-compute charm not
setting the value.  This has been removed.  Now that the function only
needs to be set for the relation, it is no longer included in 'unit'
loops.

The utility function is also renamed to
set_region_on_relation_from_config to better reflect it's actual
function.

Change-Id: I81d9924bebe4009119505b1d5dccf2e498925c7e
Related-Bug: #1833420