Patch DNS lookups in Unit tests
Currently the unit testing suite has a minimal set of IPs
hard-coded (in the subnet 10.0.0.0/8). The failing unit tests call
functions that will perform DNS reverse lookups on the hard-coded
IPs. In the event that a DNS server is unavailable, or misconfigured,
the DNS lookups will result in a Python Exception being thrown,
causing the unit tests to fail.

This solution patches DNS lookups to return None result.

Also add OS_* to tox passenv to resolve the http proxy issue in
retrieving the cirros image.

Change-Id: I7a519fbe7fba5f3aefd22c57b8d44e3b39c99582
Closes-Bug: #1841806