Explicitly unset console_access_protocol
If console_access_protocol was switched to none then
console_access_protocol was not set at all by the relation set
acting on the relation with the compute nodes. This meant the
operator had no mechanism to unset console_access_protocol. This
change explicitly unsets it so the change is passed to the
compute units.

Change-Id: Ifdca0ec6626732ee0ea30a6847130ea97478df6f
Closes-Bug: #1844356