Ignore ssh-keyscan errors on hosts
The linked bug is due to a computed shorthost name that is taken as the
first 'word' in a dotted fqdn provided by rDNS when resolving the IP
address of a nova-compute host.  However, depending on the DNS server,
this shorthost may not actually exist in the DNS server.  This patchset
ignores ssh-keyscans where effectively the host can't be reached and so
the shorthost is not added.

Change-Id: I9ca82e00a79a3eecf37231620f648683edc0bc95
Closes-Bug: #1849501