Patch basic_deploymeny.py for juju 2.7 compat
Essentially, Juju 2.7 changed the semantics of "juju run" in a way that
breaks Amulet.  Amulet is no longer supported, and this charm will have
its tests migrated to zaza in the near future.  This patch just
monkey-patches the file_contents() method of Amulet's Sentry() class so
that the memcache verify test can pass.

Change-Id: I046d9ee5fda4ca15a5824650d8aceb5d24741d94