Add new config option to set [spice].agent_enabled
This new config called spice-agent-enabled is added to inform the spice agent
is not installed in the guest instance, which in conjunction with an image
property hw_pointer_model=usbtablet allows accurate position of the mouse in
Windows guests.

This option is not rendered in the nova.conf local to nova-cloud-controller
units, instead it's passed to related nova-compute units.

Change-Id: Id64699b6d04aa05935b31d55532df45c6d973fa7
Depends-On: https://review.opendev.org/699461
Closes-Bug: #1856602