Receive dns-domain over neutron-api relation
In commit e245c0bad4c167847f79fa8c1b3099146cdaec79 we updated the Rocky
template only. We have Queens deployments that also requires this
functionality. Pike is the extant template to affect Queens.

In order to have nova-metadata-api use the same dns-domain as neutron
when nova-metadata-api runs on nova-cloud-controller we must recieve
the dns-domain on the neutron-api relation.

Change-Id: I9364cf57664695b08411cc3c6a101e132ea58cc7
Partial-Bug: #1805645