Do not mask services on db departed
Currently the charm masks all services on db departed. The consequence
of which is that when a db relation is re-joined these services are not
started back up.

In addition, it stops all services, including memcached and haproxy,
which also do not get restarted on a db re-join.

This change selectively stops but does not mask services that pertain
directly to nova. So that on db joined the correct services get started.

Change-Id: I81f59c97b33edd5c3e67c379cfdee8f26509075a