Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: I76332eb6f80b494015d461cf968604fe795b16f1
(cherry picked from commit 636c9e35985f3a462adf5c797656118c895e065f)